move multi-channel modules to separate file
